codes/requirements.txt
fastapi
uvicorn[standard]
apscheduler
beautifulsoup4
mysql-connector-python
python-dotenv==1.0.0
requests==2.31.0
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
pinecone-client==2.2.4
tiktoken==0.5.1